feat: SPH-223 list shop
